requirements.txt
streamlit==1.32.0
plotly==5.20.0
pandas==2.2.1
torch==2.2.1+cpu
transformers>=4.35.0
accelerate>=0.28.0
sentencepiece>=0.1.99
protobuf>=3
newspaper3k
SpeechRecognition
langdetect